// Chillpan/Week10/movie-list-opt/src/components/MovieCard.tsx
import type { Movie } from "../types/movie";

interface MovieCardProps {
  movie: Movie;
  onClick?: (movie: Movie) => void;
}

const MovieCard = ({ movie, onClick }: MovieCardProps) => {
  const imageBaseUrl = "https://image.tmdb.org/t/p/w500";
  const fallbackImageImage = "http://via.placeholder.com/640x480";

  const handleClick = () => {
    onClick?.(movie);
  };

  return (
    <div
      className="cursor-pointer overflow-hidden rounded-lg bg-white shadow-md transition-all hover:shadow-lg w-full"
      onClick={handleClick}
    >
      <div className="relative aspect-[2/3] overflow-hidden">
        <img
          src={
            movie.poster_path
              ? `${imageBaseUrl}${movie.poster_path}`
              : fallbackImageImage
          }
          alt={`${movie.title} 포스터`}
          className="h-full w-full object-cover transition-transform duration-300 ease-in-out hover:scale-105"
        />
        <div className="absolute right-2 top-2 rounded-md bg-black-500 px-2 py-1 text-sm font-bold text-white">
          {movie.vote_average.toFixed(1)}
        </div>
      </div>
      <div className="p-2 text-center text-sm font-semibold">
        {movie.title}
        {movie.release_date}
        {movie.original_title}
        {movie.overview}
      </div>
    </div>
  );
};

export default MovieCard;
